python/requirements.txt
rembg[gpu]==2.0.57
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
onnxruntime==1.17.3
